+++ enma_pe/pe_string_finder.cpp
@@ -3,10 +3,10 @@
 
 
 bool is_ascii(char symbol) {//eng
-	return ((unsigned int)symbol >= 0x20 && (unsigned int)symbol <= 0x7E);
+	return ((uint32_t)symbol >= 0x20 && (uint32_t)symbol <= 0x7E);
 }
 bool is_ascii_mod(char symbol) { //cyrillic
-	return ((unsigned int)symbol >= 0x20 && (unsigned int)symbol <= 0x3F) || ((unsigned int)symbol >= 0xC0 && (unsigned int)symbol <= 0xFF);
+	return ((uint32_t)symbol >= 0x20 && (uint32_t)symbol <= 0x3F) || ((uint32_t)symbol >= 0xC0 && (uint32_t)symbol <= 0xFF);
 }
 
 #define MIN_STRING_SIZE 4
@@ -31,26 +31,26 @@ void get_strings_from_image(const pe_image& image, string_base_table& string_tab
 
 
 
-	for (unsigned int section_idx = 0; section_idx < tide_image.get_sections_number(); section_idx++) {
+	for (size_t section_idx = 0; section_idx < tide_image.get_sections_number(); section_idx++) {
 	
 		pe_section * p_section = tide_image.get_section_by_idx(section_idx);
 
 		if (p_section && (!p_section->is_executable() || find_in_execute_sections)) {
 
 			BYTE * section_raw = p_section->get_section_data().data();
-			unsigned int raw_size = p_section->get_size_of_raw_data();
+            size_t raw_size = p_section->get_size_of_raw_data();
 
-			unsigned int top_ansi = 0;
-			unsigned int top_wide = 0;
-			unsigned int top_mod_ansi = 0;
-			unsigned int top_mod_wide = 0;
+            size_t top_ansi = 0;
+            size_t top_wide = 0;
+            size_t top_mod_ansi = 0;
+            size_t top_mod_wide = 0;
 
-			for (unsigned int raw_idx = 0; raw_idx < raw_size;) {
+			for (size_t raw_idx = 0; raw_idx < raw_size;) {
 				
 				if (is_ascii(section_raw[raw_idx])) {
 					if (top_wide <= raw_idx && raw_idx + 1 < raw_size && (section_raw[raw_idx + 1] == 0)) {
 
-						unsigned int i = 0;
+                        size_t i = 0;
 						for (i = 0; 
 							(raw_idx + i) < raw_size &&
 							i / 2 < MAX_STRING_SIZE && is_ascii(section_raw[(raw_idx + i)]) &&
@@ -74,7 +74,7 @@ void get_strings_from_image(const pe_image& image, string_base_table& string_tab
 					}
 					else {
 						if (top_ansi <= raw_idx) {
-							unsigned int i = 0;
+                            size_t i = 0;
 							for (i = 0;
 								(raw_idx + i) < raw_size && i < MAX_STRING_SIZE && is_ascii(section_raw[(raw_idx + i)]);
 								i++
@@ -100,7 +100,7 @@ void get_strings_from_image(const pe_image& image, string_base_table& string_tab
 				if (is_ascii_mod(section_raw[raw_idx])) {
 					if (top_mod_wide <= raw_idx && raw_idx + 1 < raw_size && (section_raw[raw_idx + 1] == 0)) {
 
-						unsigned int i = 0;
+                        size_t i = 0;
 						for (i = 0;
 							(raw_idx + i) < raw_size &&
 							i / 2 < MAX_STRING_SIZE && is_ascii_mod(section_raw[(raw_idx + i)]) &&
@@ -125,7 +125,7 @@ void get_strings_from_image(const pe_image& image, string_base_table& string_tab
 					}
 					else {
 						if (top_mod_ansi <= raw_idx) {
-							unsigned int i = 0;
+                            size_t i = 0;
 							for (i = 0;
 								(raw_idx + i) < raw_size && i < MAX_STRING_SIZE && is_ascii_mod(section_raw[(raw_idx + i)]);
 								i++
@@ -155,11 +155,11 @@ void get_strings_from_image(const pe_image& image, string_base_table& string_tab
 	}
 
 
-	for (unsigned int i = 0; i < string_table.ansi_base.size(); i++) {
+	for (size_t i = 0; i < string_table.ansi_base.size(); i++) {
 		bool b_erase_string = true;
 		char last_ch = 0;
 
-		for (unsigned int ch_idx = 0; ch_idx < string_table.ansi_base[i].str.size(); ch_idx++) {
+		for (size_t ch_idx = 0; ch_idx < string_table.ansi_base[i].str.size(); ch_idx++) {
 			if (!ch_idx) {last_ch = string_table.ansi_base[i].str[ch_idx];}
 			else {
 				if (last_ch != string_table.ansi_base[i].str[ch_idx]) {
@@ -175,10 +175,10 @@ void get_strings_from_image(const pe_image& image, string_base_table& string_tab
 		}
 	}
 
-	for (unsigned int i = 0; i < string_table.wide_base.size(); i++) {
+	for (size_t i = 0; i < string_table.wide_base.size(); i++) {
 		bool b_erase_string = true;
 		wchar_t last_ch = 0;
-		for (unsigned int ch_idx = 0; ch_idx < string_table.wide_base[i].str.size(); ch_idx++) {
+		for (size_t ch_idx = 0; ch_idx < string_table.wide_base[i].str.size(); ch_idx++) {
 			if (!ch_idx) { last_ch = string_table.wide_base[i].str[ch_idx]; }
 			else {
 				if (last_ch != string_table.wide_base[i].str[ch_idx]) {
